Added functionality so that primary beam correction code can take multiple files at once. Code should be backwards compatible.
